% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootStrap.R
\name{BootStrap}
\alias{BootStrap}
\title{Boot Strap}
\usage{
BootStrap(dat, deltas, alpha)
}
\arguments{
\item{dat}{gene count matrix}

\item{deltas}{deltas matrix}

\item{alpha}{user set threshold for reporting results (e.g., 0.05 returns values <= 0.05)}

\item{percent.zeros.retained}{Genes with <= percent.zeros.retained are retained}
}
\value{
An updated deltas matrix (last column has the p-values)
}
\description{
Calculate significance of each gene via permutation. Lower p-values represent less among-sample variation.
Plots "significant" genes as red points
}
