% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneStandardization.R
\name{GeneStandardization}
\alias{GeneStandardization}
\title{Gene Standardization}
\usage{
GeneStandardization(dat)
}
\arguments{
\item{dat}{gene count matrix}
}
\value{
A standarized count matrix
}
\description{
Standardize across genes with different gene lengths
Standardize by residuals between gene length and total expression; linear by default
}
