% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchBack.R
\name{MatchBack}
\alias{MatchBack}
\title{Match Back}
\usage{
MatchBack(delta, dat, ref2)
}
\arguments{
\item{delta}{delta matrix}

\item{dat}{gene count matrix}
}
\value{
returns plots for visualization purposes; no matrices returned
}
\description{
Matches back significant genes to standardized and plot results
}
