% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankOrder.R
\name{RankOrder}
\alias{RankOrder}
\title{Rank Order}
\usage{
RankOrder(dat)
}
\arguments{
\item{dat}{gene count matrix}
}
\value{
a delta matrix
}
\description{
This function calculates the rank-order expression (among samples) within a gene
}
