% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleStandardization.R
\name{SampleStandardization}
\alias{SampleStandardization}
\title{Sample Standardization}
\usage{
SampleStandardization(dat)
}
\arguments{
\item{dat}{gene count matrix}
}
\value{
A standardized count matrix
}
\description{
This function standardizes gene counts for samples (e.g., libraries, individuals) with different total numbers of gene counts
It also reports the mean, median gene counts per sample and plots the values.
}
