% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeroRemover.R
\name{ZeroRemover}
\alias{ZeroRemover}
\title{Zero Remover}
\usage{
ZeroRemover(dat, percent.zeros.retained)
}
\arguments{
\item{dat}{gene count matrix}

\item{percent.zeros.retained}{Genes with <= percent.zeros.retained are retained}
}
\value{
A filtered count matrix
}
\description{
This function removes genes that have a count of zero "0" at every individual
This function also removes genes that have a >= user-specified percentage of samples allowed to have a count of zero; applied for each gene separately
}
